pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "roe-agent-gate"
version = "0.1.0"
description = "Out-of-Band Rules of Engagement Enforcement for Agentic Security Testing"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "roe-agent-gate contributors" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Testing",
    "Typing :: Typed",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
transformers = [
    "transformers>=4.40",
    "torch>=2.0",
    # bitsandbytes is linux/cuda only; install manually on other platforms
    "bitsandbytes>=0.43; sys_platform == 'linux'",
]
llama-cpp = [
    "llama-cpp-python>=0.2.50",
]
anthropic = [
    "anthropic>=0.40.0",
]
openai = [
    "openai>=1.30.0",
]
claude-agent-sdk = [
    "claude-agent-sdk>=0.1.0",
]
ed25519 = [
    "cryptography>=41.0",
]
all-providers = [
    "roe-agent-gate[transformers]",
    "roe-agent-gate[llama-cpp]",
    "roe-agent-gate[anthropic]",
    "roe-agent-gate[openai]",
    "roe-agent-gate[claude-agent-sdk]",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]

[project.scripts]
roe-gate = "src.__main__:main"

[project.urls]
Homepage = "https://roegate.io"
Repository = "https://github.com/Grey-Line-Interactive/ROEGATE"
Issues = "https://github.com/Grey-Line-Interactive/ROEGATE/issues"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "SIM", "TCH"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
